policy auth in controller

+++ workopia/app/Http/Controllers/JobController.php
@@ -5,13 +5,15 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Illuminate\View\View;
 use Illuminate\Http\RedirectResponse;
-use App\Models\Job;
 use Illuminate\Support\Facades\Redirect;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
+use App\Models\Job;
 
 class JobController extends Controller
 {
+    use AuthorizesRequests;
     // @desc Show all job listings
     // @route GET /jobs
     public function index(): View
@@ -82,6 +84,8 @@ class JobController extends Controller
     // @route GET /jobs/{id}/edit
     public function edit(Job $job): View
     {
+        // Check if the user is authorized to update the job
+        $this->authorize('update', $job);
         return view('jobs.edit')->with('job', $job);
     }
 
@@ -89,6 +93,10 @@ class JobController extends Controller
     // @route PUT /jobs/{id}
     public function update(Request $request, Job $job): string
     {
+        // Check if the user is authorized to update the job
+        $this->authorize('update', $job);
+
+        // Validate the request data
         $validatedData = $request->validate([
             'title' => 'required|string|max:255',
             'description' => 'required|string',
@@ -131,6 +139,10 @@ class JobController extends Controller
     // @route DELETE /jobs/{id}
     public function destroy(Job $job): RedirectResponse
     {
+
+        // Check if the user is authorized to delete the job
+        $this->authorize('delete', $job);
+
         // Delete the logo if it exists
         if ($job->company_logo) {
             Storage::delete('public/logos/' . $job->company_logo);
